backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "BillShield FastAPI backend service"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.1",
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-nvidia-ai-endpoints>=0.3.0",
    "supabase>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
